pyproject.toml
[project]
name = "esp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "esptool>=5.1.0",
    "matplotlib>=3.10.7",
    "mpremote>=1.26.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyserial>=3.5",
    "scipy>=1.16.3",
]
